Changed the counter functions like n_vertices,n_edges,... from uint to size_t
git-svn-id: http://openmesh.org/svnrepo/OpenMesh@542 fdac6126-5c0c-442c-9429-916003d36597

+++ trunk/src/OpenMesh/Core/Mesh/ArrayKernel.hh
@@ -265,10 +265,10 @@ public:
   void clear();
 
   // --- number of items ---
-  uint n_vertices()  const { return vertices_.size(); }
-  uint n_halfedges() const { return 2*edges_.size(); }
-  uint n_edges()     const { return edges_.size(); }
-  uint n_faces()     const { return faces_.size(); }
+  size_t n_vertices()  const { return vertices_.size(); }
+  size_t n_halfedges() const { return 2*edges_.size(); }
+  size_t n_edges()     const { return edges_.size(); }
+  size_t n_faces()     const { return faces_.size(); }
 
   bool vertices_empty()  const { return vertices_.empty(); }
   bool halfedges_empty() const { return edges_.empty(); }
